+++ packages/manager/apps/hub/src/__mocks__/roadmap.ts
@@ -1,0 +1,10 @@
+import { RoadmapChangelogItem } from '@/types/roadmapchangelog.type';
+
+export const roadmapItem: RoadmapChangelogItem = {
+  title: 'title',
+  url: 'url',
+  product: 'product',
+  releaseDate: 'releaseDate',
+  status: 'status',
+  changelog: 'changelog',
+};