+++ packages/manager/apps/okms/src/common/utils/location/continents.ts
@@ -1,5 +1,5 @@
-import { GeographyCode } from '@/common/types/location.type';
 import { ContinentCode } from '@/common/types/continents.type';
+import { GeographyCode } from '@/common/types/location.type';
 
 const GEOGRAPHY_CODE_TO_CONTINENT_CODE: Record<GeographyCode, ContinentCode> = {
   eu: 'EUROPE',
@@ -9,11 +9,6 @@ const GEOGRAPHY_CODE_TO_CONTINENT_CODE: Record<GeographyCode, ContinentCode> = {
   af: 'AFRICA',
 };
 
-export const getContinentCodeFromGeographyCode = (
-  geographyCode: GeographyCode,
-): ContinentCode => {
-  if (Object.keys(GEOGRAPHY_CODE_TO_CONTINENT_CODE).includes(geographyCode)) {
-    return GEOGRAPHY_CODE_TO_CONTINENT_CODE[geographyCode];
-  }
-  return 'OTHERS';
+export const getContinentCodeFromGeographyCode = (geographyCode: GeographyCode): ContinentCode => {
+  return GEOGRAPHY_CODE_TO_CONTINENT_CODE[geographyCode] ?? 'OTHERS';
 };