+++ packages/manager/apps/okms/src/common/utils/tests/init.i18n.ts
@@ -1,23 +1,26 @@
 import i18next, { i18n } from 'i18next';
+
 import { NAMESPACES } from '@ovh-ux/manager-common-translations';
 import commonActions from '@ovh-ux/manager-common-translations/public/translations/actions/Messages_fr_FR.json';
 import commonDashboard from '@ovh-ux/manager-common-translations/public/translations/dashboard/Messages_fr_FR.json';
-import commonForm from '@ovh-ux/manager-common-translations/public/translations/form/Messages_fr_FR.json';
-import commonStatus from '@ovh-ux/manager-common-translations/public/translations/status/Messages_fr_FR.json';
 import commonError from '@ovh-ux/manager-common-translations/public/translations/error/Messages_fr_FR.json';
+import commonForm from '@ovh-ux/manager-common-translations/public/translations/form/Messages_fr_FR.json';
+import commonOnboarding from '@ovh-ux/manager-common-translations/public/translations/onboarding/Messages_fr_FR.json';
 import commonRegion from '@ovh-ux/manager-common-translations/public/translations/region/Messages_fr_FR.json';
+import commonStatus from '@ovh-ux/manager-common-translations/public/translations/status/Messages_fr_FR.json';
+
+import commonOkms from '../../../../public/translations/common/Messages_fr_FR.json';
 import kmsCommon from '../../../../public/translations/key-management-service/common/Messages_fr_FR.json';
 import create from '../../../../public/translations/key-management-service/create/Messages_fr_FR.json';
+import credentials from '../../../../public/translations/key-management-service/credential/Messages_fr_FR.json';
 import dashboard from '../../../../public/translations/key-management-service/dashboard/Messages_fr_FR.json';
 import error from '../../../../public/translations/key-management-service/error/Messages_fr_FR.json';
 import guide from '../../../../public/translations/key-management-service/guide/Messages_fr_FR.json';
 import listing from '../../../../public/translations/key-management-service/listing/Messages_fr_FR.json';
 import onboarding from '../../../../public/translations/key-management-service/onboarding/Messages_fr_FR.json';
 import serviceKeys from '../../../../public/translations/key-management-service/serviceKeys/Messages_fr_FR.json';
-import credentials from '../../../../public/translations/key-management-service/credential/Messages_fr_FR.json';
 import terminate from '../../../../public/translations/key-management-service/terminate/Messages_fr_FR.json';
 import secretManager from '../../../../public/translations/secret-manager/Messages_fr_FR.json';
-import commonOkms from '../../../../public/translations/common/Messages_fr_FR.json';
 
 export const defaultLocale = 'fr_FR';
 export const defaultAvailableLocales = [defaultLocale];
@@ -31,24 +34,13 @@ function addTranslations() {
     .addResources(defaultLocale, 'key-management-service/guide', guide)
     .addResources(defaultLocale, 'key-management-service/listing', listing)
     .addResources(defaultLocale, 'key-management-service/terminate', terminate)
-    .addResources(
-      defaultLocale,
-      'key-management-service/onboarding',
-      onboarding,
-    )
-    .addResources(
-      defaultLocale,
-      'key-management-service/serviceKeys',
-      serviceKeys,
-    )
-    .addResources(
-      defaultLocale,
-      'key-management-service/credential',
-      credentials,
-    )
+    .addResources(defaultLocale, 'key-management-service/onboarding', onboarding)
+    .addResources(defaultLocale, 'key-management-service/serviceKeys', serviceKeys)
+    .addResources(defaultLocale, 'key-management-service/credential', credentials)
     .addResources(defaultLocale, NAMESPACES.ACTIONS, commonActions)
     .addResources(defaultLocale, NAMESPACES.DASHBOARD, commonDashboard)
     .addResources(defaultLocale, NAMESPACES.FORM, commonForm)
+    .addResources(defaultLocale, NAMESPACES.ONBOARDING, commonOnboarding)
     .addResources(defaultLocale, NAMESPACES.STATUS, commonStatus)
     .addResources(defaultLocale, NAMESPACES.ERROR, commonError)
     .addResources(defaultLocale, NAMESPACES.REGION, commonRegion)
@@ -57,23 +49,24 @@ function addTranslations() {
     .use({
       type: 'postProcessor',
       name: 'normalize',
-      process: (value: string) =>
-        value ? value.replace(/&amp;/g, '&') : value,
+      process: (value: string) => (value ? value.replace(/&amp;/g, '&') : value),
     });
 }
 
 export const initTestI18n = () =>
   new Promise<i18n>((resolve) => {
-    i18next.init({
-      lng: defaultLocale,
-      defaultNS: 'key-management-service',
-      ns: [],
-      supportedLngs: defaultAvailableLocales,
-      postProcess: 'normalize',
-      interpolation: {
-        escapeValue: false,
-      },
-    });
+    i18next
+      .init({
+        lng: defaultLocale,
+        defaultNS: 'key-management-service',
+        ns: [],
+        supportedLngs: defaultAvailableLocales,
+        postProcess: 'normalize',
+        interpolation: {
+          escapeValue: false,
+        },
+      })
+      .catch((error) => console.error(error));
 
     if (i18next.isInitialized) {
       addTranslations();
@@ -89,6 +82,7 @@ const commonLabels = {
   actions: commonActions,
   dashboard: commonDashboard,
   form: commonForm,
+  onboarding: commonOnboarding,
   status: commonStatus,
   error: commonError,
   region: commonRegion,