+++ packages/manager/apps/okms/src/modules/key-management-service/hooks/useFormattedDate.ts
@@ -1,4 +1,5 @@
 import { useMemo } from 'react';
+
 import { useShellContext } from '@/common/hooks/useShellContext';
 
 type UseDateProps = {
@@ -10,8 +11,6 @@ export const useFormattedDate = ({ date, options }: UseDateProps): string => {
   const { environment } = useShellContext();
   const { userLocale } = environment;
   return useMemo(() => {
-    return Intl.DateTimeFormat(userLocale.replace('_', '-'), options).format(
-      date,
-    );
-  }, [userLocale, date]);
+    return Intl.DateTimeFormat(userLocale.replace('_', '-'), options).format(date);
+  }, [userLocale, date, options]);
 };