+++ packages/manager/apps/okms/src/modules/key-management-service/pages/credential/create/identities/cell/group/IdentityGroupIdentityCell.component.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
-import { DataGridTextCell } from '@ovh-ux/manager-react-components';
 import { IdentityGroup } from '@key-management-service/types/identity.type';
+
+import { DataGridTextCell } from '@ovh-ux/manager-react-components';
 
 const IdentityGroupIdentityCell = (group: IdentityGroup) => {
   return <DataGridTextCell>{group.urn}</DataGridTextCell>;