+++ packages/manager/apps/okms/src/modules/key-management-service/pages/credential/create/identities/tile/IdentitiesTileText.component.tsx
@@ -1,4 +1,5 @@
-import React, { ReactNode } from 'react';
+import { ReactNode } from 'react';
+
 import { ODS_TEXT_PRESET } from '@ovhcloud/ods-components';
 import { OdsText } from '@ovhcloud/ods-components/react';