+++ packages/manager/apps/okms/src/modules/key-management-service/pages/service-key/deactivate-service-key-modal/DeactivateServiceKeyModal.page.tsx
@@ -1,28 +1,6 @@
-import React, { useState } from 'react';
-import { useParams, useNavigate } from 'react-router-dom';
-import { useTranslation } from 'react-i18next';
-import {
-  OdsButton,
-  OdsFormField,
-  OdsSelect,
-  OdsSpinner,
-  OdsText,
-  OdsModal,
-} from '@ovhcloud/ods-components/react';
-import {
-  ODS_BUTTON_VARIANT,
-  ODS_SPINNER_SIZE,
-  ODS_TEXT_PRESET,
-  ODS_BUTTON_COLOR,
-  ODS_MODAL_COLOR,
-} from '@ovhcloud/ods-components';
-import { useNotifications } from '@ovh-ux/manager-react-components';
-import {
-  ButtonType,
-  PageLocation,
-  PageType,
-  useOvhTracking,
-} from '@ovh-ux/manager-react-shell-client';
+import { useState } from 'react';
+
+import { useNavigate } from 'react-router-dom';
 
 import { useUpdateOkmsServiceKey } from '@key-management-service/data/hooks/useUpdateOkmsServiceKey';
 import {
@@ -31,6 +9,33 @@ import {
   OkmsServiceKeyPutState,
   OkmsServiceKeyState,
 } from '@key-management-service/types/okmsServiceKey.type';
+import { useTranslation } from 'react-i18next';
+
+import {
+  ODS_BUTTON_COLOR,
+  ODS_BUTTON_VARIANT,
+  ODS_MODAL_COLOR,
+  ODS_SPINNER_SIZE,
+  ODS_TEXT_PRESET,
+} from '@ovhcloud/ods-components';
+import {
+  OdsButton,
+  OdsFormField,
+  OdsModal,
+  OdsSelect,
+  OdsSpinner,
+  OdsText,
+} from '@ovhcloud/ods-components/react';
+
+import { useNotifications } from '@ovh-ux/manager-react-components';
+import {
+  ButtonType,
+  PageLocation,
+  PageType,
+  useOvhTracking,
+} from '@ovh-ux/manager-react-shell-client';
+
+import { useRequiredParams } from '@/common/hooks/useRequiredParams';
 
 const stateByReason: {
   [deactivationReason: string]: OkmsServiceKeyPutState;
@@ -45,10 +50,7 @@ const stateByReason: {
 };
 
 export const DisableServiceKeyModal = () => {
-  const { okmsId, keyId } = useParams() as {
-    okmsId: string;
-    keyId: string;
-  };
+  const { okmsId, keyId } = useRequiredParams('okmsId', 'keyId');
   const [deactivationReason, setDeactivationReason] = useState<
     OkmsServiceKeyDeactivationReason | undefined
   >();
@@ -65,10 +67,7 @@ export const DisableServiceKeyModal = () => {
     keyId,
     onSuccess: () => {
       clearNotifications();
-      addSuccess(
-        t('key_management_service_service-keys_deactivation_success'),
-        true,
-      );
+      addSuccess(t('key_management_service_service-keys_deactivation_success'), true);
       trackPage({
         pageType: PageType.bannerSuccess,
         pageName: 'deactivate_encryption_key',
@@ -114,16 +113,14 @@ export const DisableServiceKeyModal = () => {
         {t('key_management_service_service-keys_modal_deactivation_heading')}
       </OdsText>
       {isPending ? (
-        <OdsSpinner className="block my-4" size={ODS_SPINNER_SIZE.md} />
+        <OdsSpinner className="my-4 block" size={ODS_SPINNER_SIZE.md} />
       ) : (
-        <OdsFormField className="block my-4">
+        <OdsFormField className="my-4 block">
           <OdsSelect
             name="deactivation-reason"
             value={deactivationReason}
             onOdsChange={(e) =>
-              setDeactivationReason(
-                e?.detail.value as OkmsServiceKeyDeactivationReason,
-              )
+              setDeactivationReason(e?.detail.value as OkmsServiceKeyDeactivationReason)
             }
             placeholder={t(
               'key_management_service_service-keys_modal_deactivation_reason_select_placeholder',
@@ -161,9 +158,7 @@ export const DisableServiceKeyModal = () => {
         color={ODS_BUTTON_COLOR.primary}
         onClick={handleSubmit}
         aria-label="edit-name-okms"
-        label={t(
-          'key_management_service_service-keys_deactivation_button_confirm',
-        )}
+        label={t('key_management_service_service-keys_deactivation_button_confirm')}
       />
     </OdsModal>
   );