+++ packages/manager/apps/okms/src/modules/secret-manager/components/breadcrumb/items/RootBreadcrumbItem.component.tsx
@@ -1,15 +1,17 @@
-import React from 'react';
-import { useNavigate } from 'react-router-dom';
-import { OdsBreadcrumbItem } from '@ovhcloud/ods-components/react';
+import { useHref } from 'react-router-dom';
+
 import {
   SECRET_MANAGER_ROUTES_URIS,
   SECRET_MANAGER_ROUTES_URLS,
 } from '@secret-manager/routes/routes.constants';
 import { useTranslation } from 'react-i18next';
+
+import { OdsBreadcrumbItem } from '@ovhcloud/ods-components/react';
+
 import { BREADCRUMB_ITEM_TEST_IDS } from './BreadcrumbItem.constants';
 
 export const RootBreadcrumbItem = () => {
-  const navigate = useNavigate();
+  const link = useHref(SECRET_MANAGER_ROUTES_URLS.root);
   const { t } = useTranslation('secret-manager');
 
   return (
@@ -17,8 +19,7 @@ export const RootBreadcrumbItem = () => {
       data-testid={BREADCRUMB_ITEM_TEST_IDS.ROOT}
       key={SECRET_MANAGER_ROUTES_URIS.root}
       label={t('secret_manager')}
-      onClick={() => navigate(SECRET_MANAGER_ROUTES_URLS.root)}
-      href={null}
+      href={link}
     />
   );
 };