+++ packages/manager/apps/pci-workflow/src/types/opaque.d.ts
@@ -1,0 +1,12 @@
+declare const base: unique symbol;
+
+declare const brand: unique symbol;
+
+declare type Opaque<BaseType, BrandType = unknown> = BaseType & {
+  readonly [base]: BaseType;
+  readonly [brand]: BrandType;
+};
+
+declare type BrandType<OpaqueType extends Opaque<unknown>> = OpaqueType[typeof brand];
+
+declare type BaseType<OpaqueType extends Opaque<unknown>> = OpaqueType[typeof Symbols.base];