+++ packages/manager/apps/hpc-vmware-public-vcf-aas/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.26.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-hpc-vmware-public-vcf-aas-app@0.25.22...@ovh-ux/manager-hpc-vmware-public-vcf-aas-app@0.26.0) (2025-12-05)
+
+
+### Bug Fixes
+
+* hpc vmware public tests config ([08606f3](https://github.com/ovh/manager/commit/08606f3eca3a889a8fc2c1206b2ea83c91485be9)), closes [#MANAGER-20363](https://github.com/ovh/manager/issues/MANAGER-20363)
+* hpc vmware public tests config ([4da964f](https://github.com/ovh/manager/commit/4da964f4b1fba6a8698a290dfb338d76c6ed9c48)), closes [#MANAGER-20363](https://github.com/ovh/manager/issues/MANAGER-20363)
+
+
+### Features
+
+* use fixed versions for tests setup ([f917f0c](https://github.com/ovh/manager/commit/f917f0cb6a3c06f1cceb8a1bca8d1f42d66c4ee7)), closes [#MANAGER-20363](https://github.com/ovh/manager/issues/MANAGER-20363)
+
+
+
+
+
 ## [0.25.22](https://github.com/ovh/manager/compare/@ovh-ux/manager-hpc-vmware-public-vcf-aas-app@0.25.21...@ovh-ux/manager-hpc-vmware-public-vcf-aas-app@0.25.22) (2025-12-01)
 
 **Note:** Version bump only for package @ovh-ux/manager-hpc-vmware-public-vcf-aas-app